requirements.txt
anthropic
python-docx
rich
python-dotenv
